create new file conftest.pн (for fixtures) and added a new check for URL match
